--- src/components/Alert/Alert.scss
@@ -1,58 +1,0 @@
-$alertSuccessColor: rgb(38, 203, 124);
-$alertInfoColor: rgb(56, 132, 255);
-$alertDangerColor: rgb(255, 70, 66);
-$alertWarningColor: rgb(247, 125, 5);
-
-.Alert {
-  width: 100%;
-  background: #f5f7f9;
-  margin: 32px 0;
-  padding: 24px 24px 24px 50px;
-  position: relative;
-  border-left: 4px solid;
-  border-radius: 3px;
-  > p:last-child {
-    margin-bottom: 0;
-  }
-}
-
-.dark .Alert {
-  background: #3a3a3a;
-}
-
-.Alert--danger {
-  border-color: $alertDangerColor;
-  .Alert__icon svg {
-    stroke: $alertDangerColor;
-  }
-}
-
-.Alert--success {
-  border-color: $alertSuccessColor;
-  .Alert__icon svg {
-    stroke: $alertSuccessColor;
-  }
-}
-
-.Alert--info {
-  border-color: $alertInfoColor;
-  .Alert__icon svg {
-    stroke: $alertInfoColor;
-  }
-}
-
-.Alert--warning {
-  border-color: $alertWarningColor;
-  .Alert__icon svg {
-    stroke: $alertWarningColor;
-  }
-}
-
-.Alert__icon {
-  top: 24px;
-  left: 16px;
-  margin: 0px;
-  display: block;
-  padding: 0px;
-  position: absolute;
-}
